+++ src/functions/writeHeaders.ts
@@ -5,25 +5,16 @@ import Status from "@/enums/status"
 import parseContent from "@/functions/parseContent"
 import { isDate } from "util/types"
 import toString from "@/functions/toString"
+import { STATUS_CODES } from "http"
 
 /**
  * Write HTTP Headers
  * @since 9.0.0
 */ export default async function writeHeaders(body: ArrayBuffer | null, context: RequestContext, rawContext: HttpContext): Promise<boolean> {
-	if (context.global.options.performance.eTag && body) {
-		const eTag = toETag(body, context.response.headers.toJSON(), context.response.cookies.toJSON(), context.response.status)
-
-		if (eTag) {
-			if (context.headers.get('if-none-match') === eTag) {
-				rawContext.status(Status.NOT_MODIFIED, 'Not Modified').write(new ArrayBuffer(0))
-				return false
-			}
-
-			context.response.headers.set('etag', eTag)
-		}
-	}
-
 	for (const [ key, value ] of context.response.headers) {
+		if (key === 'rjweb-server' || key === 'date') continue
+		if (rawContext.getCompression() && key === 'content-length') continue
+
 		rawContext.header(key, value)
 	}
 
@@ -42,7 +33,32 @@ import toString from "@/functions/toString"
 		rawContext.header('set-cookie', cookieString)
 	}
 
-	if (context.vary.size) rawContext.header('vary', Array.from(context.vary).join(', '))
+	if (context.vary.size && !context.response.headers.has('rjweb-server')) rawContext.header('vary', Array.from(context.vary).join(', '))
+
+	if (context.global.options.performance.eTag && body) {
+		const eTag = toETag(body, context.response.headers.json(), context.response.cookies.json(), context.response.status)
+
+		if (eTag) {
+			if (context.headers.get('if-none-match') === eTag) {
+				await rawContext.status(Status.NOT_MODIFIED, 'Not Modified').write(new ArrayBuffer(0))
+
+				return false
+			}
+
+			context.response.headers.set('etag', eTag)
+		}
+	}
+
+	if (context.url.method === 'HEAD' && context.global.options.methods.head) {
+		if (!context.response.headers.has('content-length') && body && !rawContext.getCompression()) rawContext.header('content-length', body.byteLength.toString())
+		else if ((!body && !context.response.headers.has('content-length')) || rawContext.getCompression()) rawContext.header('transfer-encoding', 'chunked')
+
+		await rawContext
+			.status(context.response.status, context.response.statusText || STATUS_CODES[context.response.status] || 'Unknown')
+			.write(new ArrayBuffer(0))
+
+		return false
+	}
 
 	return true
 }